+++ src/bcplncov.c
@@ -348,7 +348,7 @@ void plgnderghi(int n, int m, double **alp, double *b, int r,
 #ifndef R
   fn=(double *) malloc((np+1) * sizeof(double));
 #else
-  fn=(double *) Calloc((np+1),double);
+  fn=(double *) calloc((np+1),sizeof(double));
 #endif
   if(ideriv==1) nd=np; else nd=0;
   // loop for probabilities and partial derivatives
@@ -452,8 +452,8 @@ void Rbclcov(int *n00, int *m0, int *nobs0, double *param, double *Vout, int *nq
 
   n0= *n00; m=*m0; nobs=*nobs0;
   //if(*iprint==1) Rprintf("\nnq=%d\n",*nq);
-  x=(double *) Calloc((*nq+1), double);
-  w=(double *) Calloc((*nq+1), double);
+  x=(double *) calloc((*nq+1), sizeof(double));
+  w=(double *) calloc((*nq+1), sizeof(double));
   gauher(x,w,*nq);
   for (j=1;j<=*nq;j++) x[j]*=M_SQRT2;
   for (j=1;j<=*nq;j++) w[j]/=SQRTPI;
@@ -467,12 +467,12 @@ void Rbclcov(int *n00, int *m0, int *nobs0, double *param, double *Vout, int *nq
   MM=dmatrix(np+1,np+1);
   D=dmatrix(np+1,np+1);
   derjk=gmatrix(n0*(n0-1)/2+1,(m*m)+1,np+1);
-  der=(double *) Calloc((np+1), double);
-  cb=(int *) Calloc((n0+2), int);
+  der=(double *) calloc((np+1), sizeof(double));
+  cb=(int *) calloc((n0+2), sizeof(int));
   for(i=1;i<=n0;i++) cb[i]= (i*(i-1))/2;
   ip=0;
   alp=dmatrix(n0+1,m);
-  b=(double *) Calloc((n0+1), double);
+  b=(double *) calloc((n0+1), sizeof(double));
   ip=0; // ip++ after param later
   for(i=1;i<=n0;i++)
   { for(j=1;j<m;j++)