+++ src/m2.c
@@ -143,12 +143,12 @@ double m2stat(int n, int m, int ng, double *pp, double **alp, double *b,
   e=(double *) malloc((nm+1) * sizeof(double));
   rvec=(int *) malloc((nm+1) * sizeof(int));
 #else
-  ii=(int *) Calloc((n+1), int);
-  kk=(int *) Calloc((n+1), int);
+  ii=(int *) calloc((n+1), sizeof(int));
+  kk=(int *) calloc((n+1), sizeof(int));
   del=dmatrix(nm+1,np+1);
-  pr=(double *) Calloc((nm+1), double);
-  e=(double *) Calloc((nm+1), double);
-  rvec=(int *) Calloc((nm+1), int);   
+  pr=(double *) calloc((nm+1), sizeof(double));
+  e=(double *) calloc((nm+1), sizeof(double));
+  rvec=(int *) calloc((nm+1), sizeof(int));   
 #endif
 
   iimat=imatrix(nm+1,2);
@@ -206,10 +206,10 @@ double m2stat(int n, int m, int ng, double *pp, double **alp, double *b,
   ss=(int *) malloc((n+1) * sizeof(int));
   tt=(int *) malloc((n+1) * sizeof(int));
 #else
-  cc=(int *) Calloc((n+1), int);
-  dd=(int *) Calloc((n+1), int);
-  ss=(int *) Calloc((n+1), int);
-  tt=(int *) Calloc((n+1), int);
+  cc=(int *) calloc((n+1), sizeof(int));
+  dd=(int *) calloc((n+1), sizeof(int));
+  ss=(int *) calloc((n+1), sizeof(int));
+  tt=(int *) calloc((n+1), sizeof(int));
 #endif
 
   /* M2: get the covariance matrix */
@@ -326,7 +326,7 @@ void plgndergh(int n, int m, double **alp, double *b, int r,
 #ifndef R
   fn=(double *) malloc((np+1) * sizeof(double));
 #else
-  fn=(double *) Calloc((np+1), double);
+  fn=(double *) calloc((np+1), sizeof(double));
 #endif
   /* loop for probabilities and partial derivatives */
   for(id=0;id<=np;id++) der[id]=0.;
@@ -403,7 +403,7 @@ void summ2fr(int n, int m, int nn, int nrec, double **dat, double *fr,
 #ifndef R
   s=(int *) malloc(m * sizeof(int));
 #else
-  s=(int *) Calloc(m, int);
+  s=(int *) calloc(m, sizeof(int));
 #endif
   s2=imatrix(m,m);
   /* univariate margins */
@@ -496,7 +496,7 @@ void Rm2 (int *nitem, int *ncateg, int *nrec, double *dataset, double *alphas,
    np=(*ncateg)*(*nitem); 
 
    dat=dmatrix(*nrec,*nitem); 
-   fr=(double *) Calloc((*nrec), double); 
+   fr=(double *) calloc((*nrec), sizeof(double)); 
 
    for(i=0,nn=0;i<(*nrec);i++) 
    { for(j=0;j<(*nitem);j++) dat[i][j] = *(dataset + (j*(*nrec)+i));  
@@ -510,7 +510,7 @@ void Rm2 (int *nitem, int *ncateg, int *nrec, double *dataset, double *alphas,
    /* printf("sample size=%d\n", nn); */
 
    nm=(*nitem)*((*ncateg)-1) + ((*nitem)*((*nitem)-1)*((*ncateg)-1)*	((*ncateg)-1))/2; 
-   pp=(double *) Calloc((nm+1), double); 
+   pp=(double *) calloc((nm+1), sizeof(double)); 
    summ2fr(*nitem,*ncateg,nn,*nrec,dat,fr,pp,&ng);  
    /* print out column vector of sample moments */
    /*
@@ -520,8 +520,8 @@ void Rm2 (int *nitem, int *ncateg, int *nrec, double *dataset, double *alphas,
    printf("\n"); */
 
    /*printf("\nnq=%d\n",nq);*/
-   x=(double *) Calloc((*nq+1), double); 
-   w=(double *) Calloc((*nq+1), double); 
+   x=(double *) calloc((*nq+1), sizeof(double)); 
+   w=(double *) calloc((*nq+1), sizeof(double)); 
    gauher(x,w,*nq); 
    for (j=1;j<=*nq;j++) x[j]*=M_SQRT2; 
    for (j=1;j<=*nq;j++) w[j]/=SQRTPI; 
@@ -537,7 +537,7 @@ void Rm2 (int *nitem, int *ncateg, int *nrec, double *dataset, double *alphas,
      } 
    } 
 
-   b=(double *) Calloc(((*nitem)+1), double); 
+   b=(double *) calloc(((*nitem)+1), sizeof(double)); 
    for(i=0;i<(*nitem);i++)  
    { b[i+1]=*(bvec + i); 
      /*printf("beta(%d)=%f\n",i,b[i]);*/
@@ -549,7 +549,7 @@ void Rm2 (int *nitem, int *ncateg, int *nrec, double *dataset, double *alphas,
    for(i=1;i<=n;i++) printf("%f ", b[i]); printf("\n");*/
    np=(*nitem)*(*ncateg); 
 
-   c2=(int *) Calloc(((*nitem)+2), int); 
+   c2=(int *) calloc(((*nitem)+2), sizeof(int)); 
    /* x2=x2statb(n,m,dat,fr,alp,b,nn,x,w,nq); */
    x2=x2statb(*nitem,*ncateg,*nrec,dat,fr,alp,b,nn,x,w,*nq); 
    /*printf("\nx2=%f, df=%e\n", x2,n2-np-1);*/