+++ src/nrmin.c
@@ -201,8 +201,8 @@ double **dmatrix(int nrow, int ncol)
   avec=(double *)malloc((unsigned) (nrow*ncol)*sizeof(double));
   amat=(double **)malloc((unsigned) nrow*sizeof(double*));
 #else
-  avec=(double *)Calloc((unsigned) (nrow*ncol), double);
-  amat=(double **)Calloc((unsigned) nrow, double*);    
+  avec=(double *)calloc((unsigned) (nrow*ncol), sizeof(double));
+  amat=(double **)calloc((unsigned) nrow, sizeof(double*));    
 #endif
   for(i=0;i<nrow;i++) amat[i]=avec+i*ncol;
   return amat;